Various code clean ups to support a clean build VC7
git-svn-id: https://svn.apache.org/repos/asf/xml/security/trunk@350989 13f79535-47bb-0310-9956-ffa450edef68

+++ c/src/canon/XSECC14n20010315.cpp
@@ -192,7 +192,7 @@ bool visiblyUtilises(DOMNode *node, safeBuffer &ns) {
 
 bool XSECC14n20010315::inNonExclNSList(safeBuffer &ns) {
 
-	int size = m_exclNSList.size();
+	int size = (int) m_exclNSList.size();
 
 	for (int i = 0; i < size; ++i) {
 
@@ -342,7 +342,7 @@ XSECC14n20010315::~XSECC14n20010315() {
 		delete mp_formatter;
 
 	// Clear out the exclusive namespace list
-	int size = m_exclNSList.size();
+	int size = (int) m_exclNSList.size();
 
 	for (int i = 0; i < size; ++i) {
 
@@ -446,7 +446,7 @@ int XSECC14n20010315::XPathSelectNodes(const char * XPathExpr) {
 	XercesDocumentWrapper *theWrapper = theParserLiaison.mapDocumentToWrapper(theDoc);
 	XercesWrapperNavigator theWrapperNavigator(theWrapper);
 
-	int size = theResult.getLength();
+	int size = (int) theResult.getLength();
 	const DOMNode *item;
 	
 	for (int i = 0; i < size; ++ i) {
